+++ libimg/float_image_gradient.c
@@ -1,5 +1,5 @@
 /* See {float_image_gradient.h}. */
-/* Last edited on 2023-11-26 06:45:28 by stolfi */
+/* Last edited on 2024-12-05 00:48:56 by stolfi */
 
 #include <math.h>
 #include <assert.h>
@@ -17,11 +17,11 @@
 
 /* IMPLEMENTATIONS */
 
-void float_image_gradient_sobel(float_image_t *A, int cA, float_image_t *DX, int cX, float_image_t *DY, int cY)
+void float_image_gradient_sobel(float_image_t *A, int32_t cA, float_image_t *DX, int32_t cX, float_image_t *DY, int32_t cY)
   {
     /* Get the image dimensions: */
-    int NX = (int)A->sz[1]; 
-    int NY = (int)A->sz[2];
+    int32_t NX = (int32_t)A->sz[1]; 
+    int32_t NY = (int32_t)A->sz[2];
     demand((cA >= 0) && (cA < A->sz[0]), "invalid {A} channel");
     
     bool_t has_DX = ((cX >= 0) && (DX != NULL));
@@ -39,7 +39,7 @@ void float_image_gradient_sobel(float_image_t *A, int cA, float_image_t *DX, int
     if ((! has_DX) && (! has_DY)) { return; }
     
     /* Fill images: */
-    int x, y;
+    int32_t x, y;
     for (x = 0; x < NX; x++)
       { for (y = 0; y < NY; y++)
           { /* Compute the horizontal and vertical derivatives in channel {c}: */
@@ -52,19 +52,19 @@ void float_image_gradient_sobel(float_image_t *A, int cA, float_image_t *DX, int
       }
   }
 
-void float_image_gradient_sqr_sobel(float_image_t *A, int cA, float_image_t *G, int cG)
+void float_image_gradient_sqr_sobel(float_image_t *A, int32_t cA, float_image_t *G, int32_t cG)
   {
     /* Get the image dimensions: */
-    int NCA = (int)A->sz[0];
+    int32_t NCA = (int32_t)A->sz[0];
     demand((cA >= 0) && (cA < NCA), "invalid {A} channel");
-    int NCG = (int)G->sz[0];
+    int32_t NCG = (int32_t)G->sz[0];
     demand((cG >= 0) && (cG < NCG), "invalid {G} channel");
-    int NX = (int)A->sz[1]; 
-    int NY = (int)A->sz[2];
+    int32_t NX = (int32_t)A->sz[1]; 
+    int32_t NY = (int32_t)A->sz[2];
     float_image_check_size(G, -1, NX, NY);
     
     /* Fill it: */
-    int x, y;
+    int32_t x, y;
     for (x = 0; x < NX; x++)
       { for (y = 0; y < NY; y++)
           { /* Compute the horizontal and vertical derivatives in channel {c}: */
@@ -84,7 +84,7 @@ float_image_t *float_image_gradient_sqr_relative
   )
   {
     /* Get the image dimensions: */
-    int NC, NX, NY;
+    int32_t NC, NX, NY;
     float_image_get_size(A, &NC, &NX, &NY);
     
     /* Allocate the output image: */
@@ -95,21 +95,21 @@ float_image_t *float_image_gradient_sqr_relative
     float_image_t *V2 = float_image_new(1, NX, NY);
     
     /* Weight table: */
-    int hw = 2, nw = 2*hw+1;
+    int32_t hw = 2, nw = 2*hw+1;
     double wt[nw];
-    wt_table_binomial_fill(nw, wt, NULL);
-    wt_table_normalize_sum(nw, wt);
+    wt_table_binomial_fill((uint32_t)nw, wt, NULL);
+    wt_table_normalize_sum((uint32_t)nw, wt);
  
     /* Process channel by channel: */
     double noise2 = noise*noise;
     double scale = (mono ? 1.0 : 1.0/NC);
-    int c;
+    int32_t c;
     for (c = 0; c < NC; c++)
       { /* Compute the gradient squared and variance for channel {c}: */
         float_image_gradient_sqr_sobel(A, c, G2, 0);
         float_image_local_avg_var(A, c, hw, wt, NULL, 0, V2, 0);
         /* Divive one by the other, store/add in {R}: */
-        int x, y;
+        int32_t x, y;
         for (x = 0; x < NX; x++)
           { for (y = 0; y < NY; y++)
               { double g2 = float_image_get_sample(G2, 0, x, y);
