+++ libimg/float_image_to_uint16_image.c
@@ -1,5 +1,5 @@
 /* See {float_image_to_uint16_image.h} */
-/* Last edited on 2017-06-22 17:43:35 by stolfilocal */ 
+/* Last edited on 2024-12-05 01:01:56 by stolfi */ 
 
 #include <limits.h>
 #include <assert.h>
@@ -20,21 +20,21 @@
 uint16_image_t *float_image_to_uint16_image
   ( float_image_t *fim,  /* Float image to convert. */
     bool_t isMask,       /* TRUE for masks, FALSE for images. */
-    int chns,            /* Number of channels of output image. */
+    int32_t chns,            /* Number of channels of output image. */
     double lo[],         /* Nominal min float sample for each chosen channel. */
     double hi[],         /* Nominal max float sample for each chosen channel. */
-    int ch[],            /* Indices of channels of {fim} to convert. */
+    int32_t ch[],            /* Indices of channels of {fim} to convert. */
     uint16_t maxval, /* Max integer sample value in result image. */
     bool_t yup,          /* If TRUE, reverses the indexing of rows. */
     bool_t verbose       /* If TRUE, prints conversion diagnostics to {stderr}. */
   )
   { /* Get image dimensions: */
-    int NX = (int)fim->sz[1];
-    int NY = (int)fim->sz[2];
+    int32_t NX = (int32_t)fim->sz[1];
+    int32_t NY = (int32_t)fim->sz[2];
     
     /* Channel counts: */
-    int fchns = (int)fim->sz[0]; /* Num channels in float image. */
-    int ichns = chns;            /* Num channels in integer image. */
+    int32_t fchns = (int32_t)fim->sz[0]; /* Num channels in float image. */
+    int32_t ichns = chns;            /* Num channels in integer image. */
     
     /* Allocate PGM/PPM image: */
     uint16_image_t *pim = uint16_image_new(NX, NY, ichns);
@@ -43,13 +43,13 @@ uint16_image_t *float_image_to_uint16_image
     pim->maxval = maxval;
     
     /* Channel indexing variables: */
-    int k; /* Channel of integer image. */
-    int c; /* Channel of float image. */
+    int32_t k; /* Channel of integer image. */
+    int32_t c; /* Channel of float image. */
     
     /* Input and output range registers: */
     float vmin[ichns], vmax[ichns];         /* Float pixel range. */
-    sample_uint32_t imin[ichns], imax[ichns]; /* Int pixel range. */
-    int clo[ichns], chi[ichns];             /* Counts of lo-clipped and hi-clipped pixels. */
+    sample_uint32_t imin[ichns], imax[ichns]; /* Int32_T pixel range. */
+    int32_t clo[ichns], chi[ichns];             /* Counts of lo-clipped and hi-clipped pixels. */
     for (k = 0; k < ichns; k++) 
       { clo[k] = chi[k] = 0;
         vmin[k] = +INF;
@@ -59,9 +59,9 @@ uint16_image_t *float_image_to_uint16_image
       }
     
     /* Convert pixels, store in {pim}, keep statistics: */
-    int x, y;
+    int32_t x, y;
     for(y = 0; y < NY; y++)
-      { int ppmy = (yup ? NY - 1 - y : y);
+      { int32_t ppmy = (yup ? NY - 1 - y : y);
         uint16_t *prow = pim->smp[ppmy];
         for(x = 0; x < NX; x++)
           { /* Convert float pixel {fpxy[c..c+2]} to integer pixel {ipxy[0..2]}, keep stats: */
@@ -81,8 +81,8 @@ uint16_image_t *float_image_to_uint16_image
     
     if (verbose)
       { /* Print statistics: */
-        long int NPIX = ((long int)NX)*((long int)NY);
-        fprintf(stderr, "  %ld pixels in float image\n", NPIX);
+        int32_t NPIX = ((int32_t)NX)*((int32_t)NY);
+        fprintf(stderr, "  %d pixels in float image\n", NPIX);
         if (NPIX > 0)
           { for (k = 0; k < chns; k++)
               { double lok = (lo == NULL ? 0.0 : lo[k]);
